+++ Flow/Worker.ipp
@@ -86,7 +86,7 @@ inline std::uint32_t Flow::Worker::dispatchSwitchNode(Node * const node)
     auto &switchTask = std::get<static_cast<std::size_t>(NodeType::Switch)>(node->workData);
     const auto index = switchTask.func();
     const auto count = node->linkedTo.size();
-    std::size_t joinCount = 1u;
+    std::uint32_t joinCount = 1u;
 
     coreAssert(!node->bypass.load(),
         throw std::logic_error("A branch task can't be bypassed"));
@@ -95,7 +95,7 @@ inline std::uint32_t Flow::Worker::dispatchSwitchNode(Node * const node)
     coreAssert(switchTask.joinCounts.size() == count,
         throw std::logic_error("Invalid switch task preprocessing, expected " + std::to_string(count) + " join counts but have " + std::to_string(switchTask.joinCounts.size())));
     scheduleNode(node->linkedTo[index]);
-    for (std::size_t i = 0; i < count; ++i) {
+    for (std::uint32_t i = 0; i < count; ++i) {
         if (i != index)
             joinCount += switchTask.joinCounts[i];
         else
